WEB-4654 abstract out TIR range filter
